Implemented AbstractEntity and Player. Tested.

+++ core/src/org/academiadecodigo/bootcamp/simulation/entities/Movable.java
@@ -1,9 +1,7 @@
 package org.academiadecodigo.bootcamp.simulation.entities;
 
-import com.sun.javafx.scene.traversal.Direction;
-
 public interface Movable {
 
-    void move(Direction direction, int amount);
+    void move(Direction direction, double amount);
 
 }
